Health UI

+++ Soul Hero/Assets/Scripts/EnemyControler.cs
@@ -23,7 +23,7 @@ public class EnemyControler : MonoBehaviour
     private float wait_Beffore_Attack_Time = 1f;
     private float attack_Timer;
 
-    public float attack_damage = 0f;
+    public int attack_damage = 0;
 
 
     private EnamyState enamy_State;
